+++ utils/train_moe.py
@@ -172,8 +172,9 @@ class LoggingTrainer(tf.Trainer):
 class WeightedSamplerTrainer(Trainer):
     def get_train_dataloader(self):
         # 从 train_dataset 中获取 sample_weights
+        # NPU does not support float64; use float32 everywhere
         sample_weights = torch.tensor(
-            self.train_dataset.sample_weights, dtype=torch.double
+            self.train_dataset.sample_weights, dtype=torch.float32
         )
         # 用 sample_weights 构建 WeightedRandomSampler
         sampler = WeightedRandomSampler(
